@@ -3,7 +3,7 @@
     "name": "Jakub Lečbych"
   },
   "description": "React tree select component  based on [react-select](https://github.com/JedWatson/react-select#react-select)\r and [react-virtualized-select](https://github.com/bvaughn/react-virtualized-select#react-virtualized-select)",
-  "version": "0.6.10",
+  "version": "0.7.0",
   "bugs": {
     "url": "https://github.com/lecbyjak/intelligent-tree-select/issues"
   },
@@ -27,19 +27,19 @@
   },
   "main": "lib/index.js",
   "dependencies": {
-    "bootstrap": "^4.1.3",
+    "bootstrap": "^4.3.1",
     "classnames": "latest",
     "informed": "^1.10.8",
-    "react-highlight-words": "^0.11.0",
-    "react-input-autosize": "^2.2.1",
+    "react-highlight-words": "^0.16.0",
+    "react-input-autosize": "^2.2.2",
     "react-select": "^1.3.0",
-    "react-virtualized": "^9.21.0",
+    "react-virtualized": "^9.21.1",
     "react-virtualized-select": "^3.1.3",
-    "reactstrap": "^6.5.0"
+    "reactstrap": "^8.1.1"
   },
   "peerDependencies": {
-    "react": "^15.3.0 || ^16.0.0-alpha",
-    "react-dom": "^15.3.0 || ^16.0.0-alpha",
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0",
     "bootstrap": "^4.1.2"
   },
   "devDependencies": {
@@ -56,8 +56,8 @@
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
     "prop-types": "^15.6.2",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.11.0",
+    "react-dom": "^16.11.0",
     "style-loader": "^0.21.0",
     "webpack": "^4.16.2",
     "webpack-cli": "^3.1.0",
