@@ -112,15 +112,15 @@ class Select extends _react2.default.Component {
         this.state = {
             inputValue: '',
             isFocused: false,
-            isOpen: this.props.isMenuOpen,
+            isOpen: props.isMenuOpen,
             isPseudoFocused: false,
             required: false
         };
         this.input = _react2.default.createRef();
+        this._instancePrefix = `react-select-${props.instanceId || ++instanceId}-`;
     }
 
-    componentWillMount() {
-        this._instancePrefix = `react-select-${this.props.instanceId || ++instanceId}-`;
+    componentDidMount() {
         const valueArray = this.getValueArray(this.props.value);
 
         if (this.props.required) {
@@ -128,9 +128,6 @@ class Select extends _react2.default.Component {
                 required: handleRequired(valueArray[0], this.props.multi)
             });
         }
-    }
-
-    componentDidMount() {
         if (typeof this.props.autofocus !== 'undefined' && typeof console !== 'undefined') {
             console.warn('Warning: The autofocus prop has changed to autoFocus, support will be removed after react-select@1.0');
         }
@@ -139,7 +136,7 @@ class Select extends _react2.default.Component {
         }
     }
 
-    componentWillReceiveProps(nextProps) {
+    UNSAFE_componentWillReceiveProps(nextProps) {
         const valueArray = this.getValueArray(nextProps.value, nextProps);
 
         if (nextProps.required) {
